Add PQAlloc move constructor/assignment in C++11.
The new constructor and assignment operator are available only in the
case where std::shared_ptr is available, but that should be the case
in C++11.

Also retires scoped_array, which was always an unnecessary duplication
of std::vector (which does guarantee C-style array layout).

For some reason the parameter-lengths array was twice as long as it
needed to be.  ISTR this was either a workaround for an ancient bug
in libpq, already fixed; or a clever trick for saving an allocation
and using the array as two arrays, which is not now needed.  Removed.