Un-inline some result member functions.
Again this weakens the compile-time dependency between row and result.
If I can get this to the point where result can be defined before row,
then it will be possible to embed a result in row, and make it
possible to pass rows and iterators around without also keeping a
separate result object just to keep the thing alive.  Which in turn
will also make it possible to give transactions a convenience function
for "issue a query, check that it returns exactly one row, and let me
access it without having to go through the result set first."

When did yaks get so _hairy_?