Replace result inheritance with member.
pqxx::result was derived from a PQAlloc managing a result_data.  That
seemed like a good idea at the turn of the century, but it doesn't now.
It makes the codebase harder to reason about, and in this case, it
doesn't even get us the benefit of the empty-base-class optimisation.