Drop result_data.
This makes pqxx::result easier to reason about, and opens the door to
more active maintenance.  Instead of each result containing a smart
pointer to a struct containing the libpq result set, the query text,
and other items, the result now directly holds a smart pointer to the
libpq result set.  It copies the query text as a string.

Performance may suffer very slightly, because the query string is now
copied every time you copy a result object.  But, the object is smaller
and simpler.