Assume more C++11 features.
Remove checks and workarounds for compilers that don't support "final",
"override", std::isnan, std::isinf, and std::move.

This will speed up the configure process a bit, and it makes the code
a lot less ugly.  The downside is that the code now expects your
compiler to implement a lot of C++11 features.

If you have an older version of gcc or clang, as found on e.g. a
long-term-support Ubuntu version, use the option "-std=c++11" option.
Or even "-std=c++14", which is the default in newer versions of the
compilers.