Stop defining NOMINMAX for Visual Studio.
It's annoying that Visual Studio defines its own min() and max() as
preprocessor macros, but the less we interfere in the client's
compilation environment, the better.

This change removes the NOMINMAX definition from libpqxx headers, and
moves the calls that clash with Visual Studio's macros out-of-line.

Some things may still break with this change.  I don't have a copy of
Visual Studio handy to test it on.