Re-do poll()/select() support.  Add WSAPoll().
This was a lot of work, but I think it simplifies a few things.

Don't fail if sys/select.h is not available.  That was a mistake; I had
started to conflate sys/select.h presence with select() support.

On Windows, always use WSAPoll().  It's simple and very similar to
poll().  It saves us checking for features; I don't think anybody runs
the configure script on Windows.

Use Autoconf Archive for more checks, and simpler AC_CHECK_HEADERS
checks instead of custom configure code.

Remove the PQXX_ prefix from some "compiler/internal" config-item names.

Continue to avoid unnecessary checks in the configure script, as well as
unnecessary code.  If WSAPoll()/poll() is available, always use that.
Otherwise, use select() with sys/select.h if present.  Otherwise, use
select() with the older headers as found on the system.