Cast fd to SOCKET for WSAPoll().
In the data structure that we're expected to pass to WSAPoll(), the file
descriptor is not an int but (according to Visual Studio 2015) something
that would require a narrowing conversion from an int.  The type is
SOCKET, which might be an int32 or something like that.

Cast the int to SOCKET.  See if that works.