Fix underflow in reverse result iterator.
This may have been causing crashes in test82 on gcc 7.2.0.  A reverse
iterator is implemented as a regular iterator, but at one element before
the element it's actually accessing.  (This can really do your head in
because the types' ideas of "forward" and "up" are reversed.)

With a reverse iterator, rbegin() points to end() - 1.  But what if the
container is empty?  The simple, ugly fix is to allow m_index to be -1.
Bear in mind that dereferencing that iterator would be illegal anyway.