Further decouple row and field from result.
This will help with a massive chicken-and-egg problem.  Test 82 crashes
on gcc 7.2.0.  The reason seems to be that it tests row iterators (so
iterators pointing to fields) that were created on a row variable which
was a temporary.

In libpqxx 6, a row contains its own copy of its result.  (A result is
mostly a shared_ptr, so that's not as weird as it sounds.)  But a field
only contained _a pointer to_ its result.  And it points at the copy
that's in the row.  When the row variable is destroyed, POOF!  There
goes the object at which that pointer was pointing.

To fix this, we'll need to give field its own copy of the result as
well.  (Hopefully most of that stuff will be optimised out, so that
there's not too much overhead.  I'm seeing a lot of fields in the
debugger that the compiler managed to optimise away.)

But before we can make field contain a result, I had to make it
possible to compile the result class definition without compiling the
field class definition first.  Otherwise we'd have a dependency cycle.