Make field contain a copy of its result.
Fixes a crash in test82 with gcc 7.2.0.

A result will now be kept alive even if it's only referenced by a field
(or row_iterator) object.  Hopefully the compiler will be able to
optimise away the overhead in most cases; it fixes a gaping pointer
bug.  Unfortunately it also complicates including the row/field headers;
get the order wrong and things won't compile.  The safe thing to do is
simply to include the result header, and not the row/field ones.