Retire unnecessary connection capabilities.
Connections have a way of checking and reporting certain capabilities of
the underlying libpq version and database backend.  This can help bridge
version differences, so I'm keeping the mechanism.  However, all of the
capabilities that had been defined are always supported in libpqxx 6.0,
and have been since 5.0, so we can get rid of them.  There are currently
no dynamic capabilities.