Start postgres manually.
Seems we can't use systemctl inside a container, so that's not the way
to start the database.

Instead, take a page out of the old libpqxx test script and create a
local database running on a Unix domain socket.  Makes everything
easier, although it probably needs more work to get it functional.