Fixed a mismatched tag in the doc/libpqxx.xml - XML Parsing Error: mismatched tag. Expected: </chapter>. Line Number 45, Column 7 (#45)
Thanks.  I guess I didn't have all the same tools installed on the laptop where I normally test.

I'd like to replace the XML-based tutorial completely, either with headers that Doxygen processes, or with Sphinx.  It'd be great if the tutorial and the reference docs could link back and forth to each other.