include sys/time.h outside fallback select (#55)
The implementation of 'wait_fd' (inside 'connection_base.cxx') always
relies on the existence of the 'timeval' structure. In Linux, this
structure is provided by the header 'sys/time.h'. If polling or select
capabilities are detected, the structure is never included into the
source and will result in a compilation error (GCC 7.x):

    connection_base.cxx:1153:28: error: ‘{anonymous}::tv_milliseconds’ declared as an ‘inline’ variable
     inline int tv_milliseconds(timeval *tv = nullptr)
                                ^~~~~~~
    ...

The following moves the 'HAVE_SYS_TIME_H' check outside the
select-fallback case so that the header can be included no matter the
event-function feature is used.

Signed-off-by: James Knight <james.knight@rockwellcollins.com>