Put test-driver under revision control.
Looks like this script is generated by autogen.sh, not by the configure
script.  If we want users to be able to configure and build without
running autogeh.sh, we'll need to add it.  And possibly more besides.

See #69.  Not closing the issue yet, because there may be other cases
of the same problem.