Check against library-vs-headers version mismatch.
This is something that keeps happening to a few users: they build their
own libpqxx, but their own code won't link against it, and the reason
turns out to be that there's an older version of the libpqxx headers
installed on the system.  Their code was compiled against this older
version, and so won't link to the newer library binary they've built.

(The custom version doesn't have to be newer than the installed one,
but that's how it usually goes.)

When that happens now, the code will fail to link.  A stub function in
the library contains the library version in its signature, and there's
some code in the headers which will try to call that stub.  If the
headers are for a different libpqxx version than the one that the
libpqxx binary was built from, the stub for that version won't be found
and the link will fail.