Allow build to hide std::experimental::optional.
Introduces a new macro, `PQXX_HIDE_EXP_OPTIONAL`, to hide support for
`std::experimental::optional` while building code for linking against
libpqxx.  This can paper over a difference in selected C++ language
versions between the build of libpqxx itself, and the build of the
software which uses it.

Fixes #93.
Fixes #95.