Support string conversions for enums.
Addresses #59, but without compiler-specific antics.  It does use a
preprocessor macro (☹) and an intermediate struct.

In C++14 I think we might make this easier: define a generic
`string_traits` for enum types.  It takes the enum's name from a global
template variable.  To enable string conversion for an enum type, just
specialise the variable for that type, setting its name.  SFINAE will
keep the generic definition out of the way as long as the variable is
not defined for a given type.