Make `next_seq` return a position, not a struct.
We never really needed the `seq_position` type.  It's conceptually nice
to find the next glyph as a byte sequence, but all we really needed was
a way to find the one-past-end position of the current glyph.  The
caller knows the start byte, because it was one of the arguments in the
first place.