Factor out glyph scanning dispatch.
Separates the "select a scanning function for this encoding group" from
the "scan for the end of this glyph."  It lets callers hoist the lookup
of the right scanner instantiation out of their inner loop.

A nice side effect is that this uses templates rather than the
preprocessor.  It's a little awkward because template template arguments
must refer to types, never functions.  To get around that, I wrapped the
functions in structs.

I think there may also have been a bug in `find_with_encoding` when
needle is a string, if a prefix of the needle occurs at the end of the
haystack.  In that situation, the code could walk off the end of the
haystack and into never-never land.