Fix cursor query stripping for encodings.
The cursor code needs to strip any remaining semicolon (plus any
trailing whitespace that might be masking one) from its query in order
to embed it in a larger `DECLARE` statement.  This was always broken
for multibyte encodings.  Thanks to @JadeMatrix, we now have the ability
to find glyph boundaries for supported enodings, so that we can scan
through the query properly and not mistake any embedded bytes for whole
characters.

This still won't work for encodings where the semicolon does not look
like the ASCII semicolon, but I don't believe any such encoding to be
supported.