Generate CMakeLists.txt files.
Use our custom templating system to generate CMake files.

This is still horrible, in that it specifies the names of the source
files as globs, even though they don't actually contain any wildcards.
I'm sure there is a proper way of specifying the files without globs,
but I simply don't know what it is yet.