In CMake, check headers in C, not in C++.
Looks like all along, CMake was breaking, and not making it quite clear
enough!  We removed C as a language, but it turns out that checking for
an "include file" _always_ uses the C compiler.  Which wasn't
configured.  And here I was thinking that it was normal that you need
to run cmake twice in a row — what posessed it to proceed to a state
where you could "successfully" generate the build despite a compiler
being missing!?

Anyway, there's an alternative module and function to check for a C++
header.  That did the trick.