Factor out "type name" from string traits.
Most of this will have to be done later, because it requires an ABI
change.  Moving the "human-readable name for this type" functionality
into a separate traits template will make it possible to define the
string traits for the built-in types using a template, rather than a
big preprocessor macro.

It can also replace part of `namedclass`.