Harmonise some copyright notices.
Generally, I think the copyright notice in any file should have the
years in which _the product_ was published.  What I've been doing (and
mostly because I enjoyed it, which has just about run its course) is
give each file its own copyright period, so you could see when the
file was created.  I'm starting a transition to the latter.

Also, when was the product first published?  In hindsight I would
argue 2000, not 2001, given today's low threshold of what it takes to
publish source code.