In CMake, prefer CMake's config headers.
Fixes #178.

Sometimes people experienced strange build problems when building
with CMake, where libpqxx seemed to be compiled for a different C++
version than the one for which it was configured.

This seems to happen in out-of-tree builds when the source tree has
the configuration headers (`include/pqxx/config-*-*.h`), generated
by the native autoconf-based build.  The CMake build would generate
similar config headers in the build tree, but when actually compiling,
the build preferred the autoconf-generated config headers over the
CMake-generated ones.

This fixes that by reordering the include directories as they are
added.