Re-deprecated _deactivation._
Sorry for wavering on this.  It may or may not end up simplifying the code
much, but from the user's perspective, was there ever really much point to
this feature?

If nothing else, knowing that the user won't deactivate a connection will
make it easier for me to reason about reactivation avoidance and
inhibition.  Currently it does my head in.