Remove session variable cache.
Yeah, there never really was a good reason for this cache, other than
to restore variables as part of connection reactivation.  And I just
ditched connection reactivation.

This change _may_ inconvenience someone out there.  But if it does,
then they were probably already on slightly thin ice.  On the other
hand, this is removing a complication which might cause trouble for
someone as well.