Drop `lazyconnection` and `asyncconnection`.
This hurt.  But it needed to be done, or the whole connection hierarchy
would remain stuck in its present, ossified state.

How do we we deal with compatibility?  Easiest thing for now is to make
`asyncconnection` an alias for the one, single connection type.  It'll
affect performance of applications geared to use it, but probably not
too much.

For `lazyconnection`, the one real use was to create larger pools of
connections than you can really afford to have.  Replacing those with
immediate connections might cause a storm of connection attempts.  So,
better for those to fail and require an application change.