Flatten `connectionpolicy` hierarchy into 1 class.
Now that there's only one connection policy left, we only need one
connection policy class — not a hierarchy.

Next step of course is to fold it all into `connection_base`.  And, at
that point we will no longer need a connection hierarchy either.  It'll
all be one class.  Applications will still expect us to have two types
`connection` and `connection_base`, but one can be an alias for the
other.