Try CircleCI build using CMake.
The baseline C++ version is now C++17, supported by autoconf-archive.
CircleCI's build slave container does not have a sufficiently recent
version of autoconf-archive to make this work.

So, try doing the CircleCI build using CMake for now.  This means we
can't test as thoroughly, but maybe it'll get the CI back up for the
short term.