Stop deriving `stream_from` from `stream_base`.
There really wasn't any point to having a streams hierarchy.  There are
similarities in usage between the two concrete classes, but nothing that
really needs to be expressed in C++.  It just makes it harder to find
which method lives where, and follow the call chains.

Next step: do the same for `stream_to`.  Then, ditch `stream_base`!