Simplify `to_chars` calls.
Turns out the `chars_format` argument to these functions is not
mandatory after all.  Integer precision was never mandatory, and the
implementations can tell integral types from floating-point types, so I
was able to drop a lot of boilerplate.

Sadly, even on g++ 9.1, `from_chars` is still not working for
floating-point types.  ☹  The overload is only available for integral
types.