Fix `from_string` overflow checks.
There were cases where overflows in converting strings to integral types
were not detected.  This should fix that.  Probably not particularly
useful, since that code is being replaced by `from_chars`.

But since my custom code in `to_string` for integral types may still be
useful (as an optimisation beyond what `to_chars` can do), I optimised
the cases for negative numbers a bit.  The only case that's still slow
is the very lowest (negative) number which a signed type can represent,
for types of 64 bits or more.  Hopefully that case will be rare, and in
the future it'll be able to fall back on `to_chars`.

Some more template trickery went away, being replaced by `if constexpr`.