Start 6.3 branch.
If you're a software professional and you know your branching
strategies, this may strike you as lunacy.  But it's how I like to do
release branches: development is almost entirely linear with releases
coming straight from `master`, and just occasionally when I really
need to update an older version, I start a "release branch."

That may not even be the right name, but boy, does it work well for me.
You get a release branch when you need one, but only pay the cost of
one very rarely — and even then, only once the code has largely
stabilised and updates are rare.  Basically, it's a rare one-off way
to fix occasional bugs.