Build to-string conversions on `to_buf`.
Replaces most of the `to_string` machinery with the `to_buf` functions,
which were now doing most of the same things.

Moves some error-prone template instantiations from the library into the
headers.  Most of the conversion work is done in the headers now, which
will increase compilation times but hopefully also the compiler's
optimisation opportunities.

I had some trouble with the `hard_neg` template constant.  I finally
realised that some of these problems were because I was using integral
constants without specifying their type.  Not an issue in normal
run-time usage, but I think `hard_neg<-128>` will not hit the
instantiation for `-128LL`.

The ordering was a nightmare to get right for all my compilers.  The
`stringstream`-based fallback for converting floating-point numbers to
strings was still needed, but its natural type is the `std::string` it
gets from the standard library.  So, in that case, I'm not building the
conversions to `std::string` on top of the write-to-raw-buffer code, but
the other way around.

Man, I'll be happy when I have working `charconv` for floating-point
types throughout!