Major string conversion changes: `str` class. (#207)
* Working on new string conversion API.

This is not ready for use yet.  I'm continually fighting template
specialisation problems (in a truce now), and tests still fail on
clang 8 (because `std::optional<int>` does not convert properly).

* Get `to_string` working for smart pointers etc.

_Finally_ got passing tests for `to_string` with `std::optional`,
`std::unique_ptr`, and `std::shared_ptr`.

Trimmed down the `str_impl` classes a bit, moving the redundant
`operator std::string_view` to the `str` interface class.

* Make `has_null` a variable.

No reason why `has_null` should still be a function in C++17.

* Get rid of `type_utils.hxx`.

Turns out there was nothing in here that we're using at the moment.  If
we need it again, we'll have to dig it up or reconstruct it.  But I
prefer that over dead code, especially if it's this intricate!

* Obviate irrelevant `null()`  functions.

Remove `null()` from `string_traits` specialisations for which
`has_null` is `false`.  Throw the null conversion exception at the
call site.  Long live  `if constexpr`!

* Make traits `from_string` return its result.

Make the signature for `string_traits<T>::from_string` a bit more
sensible.  It now just returns a value, instead of taking an output
parameter.  The RVO should make the `return` easy.

This does leave an unpleasing asymmetry: `pqxx::from_string` takes
an output argument, which saves us having to instantiate `from_string`
explicitly every time we call it.

* Don't require `is_null` where there's no null.

A single `pqxx::is_null` now answers the question, "is this object a
null value of its type?"  It uses `if constexpr` to avoid calling the
traits type's `is_null` if the type has no null, so there's no need
to define it.