Introduce `zview` class.
This is a marker type.  It wraps a `std::string_view`, but adds the
guarantee that there be a terminating zero right after, at the "end"
position of the `string_view`.

(This is not checked anywhere.  The caller needs to make sure.  But
after that, it makes it a little safer and clearer to build code which
assumes the terminating zero.)