Lint.
Trying out `cpplint.py` from the C++ Core Guidelines:

    https://github.com/isocpp/CppCoreGuidelines

Not all of the rules in this script appeal to me, so I was highly
selective in applying them.  It would be good to have a uniform style,
though having opening braces on their own line is a must-have for me.

Done while listening to Nicolai Josuttis's CppCon 2019 talk.  He says
the script's virtual/override rule is not perfect.