Extract `nullness` traits.
The notion of extracting all string conversions to stand-alone functions
has failed.  It needed partial specialisation, which isn't allowed with
functions.

But I don't think the idea of `string_traits` evolving towards a
"null traits" type was a bad one.  So, I'm splitting `string_traits` up
into one traits type for nullness, and one for string conversions.

Next steps, as I see them now:
 1. Return `from_string` specialisations/overloads to `string_traits`.
 2. Templatise `string_traits` on text/binary format.
 3. Start implementing binary conversions.

Binary conversions are a headache, in that the format is poorly
specified and the required type-punning looks hard to get actually
right (as opposed to "WFM") in C++.  But I know there's a library out
there, so hey, maybe I can call that if it's available.