Strip whitespace from pkg-config output.
Thanks @rjzupkoii for analysing and reporting the problem, and proposing
the fix.  It's a good fix, just in the wrong place!  The `configure`
script gets automatically generated, on a regular basis, so we don't
make changes there.  Instead, fix it in `configure.ac` which is the
"source file" for `configure`.

Fixes #225.