Try exporting members, not classes.
I'm trying to get the build working on Visual Studio 2019.  It's giving
me crazy errors: "unresolved external symbol" for inline functions.

My working theory is that all members of a class inherit the export
declaration for the class — even when those members are inline
functions.  And that declaration turns into an _import_ in code which
calls the library.  You'd expect the compiler to ignore the import for
an inline function, but perhaps it doesn't.