build: autotools: use libtool -version-info for shared library SONAME
When using the libtool -release flag, every release of the package will be binary incompatible with any other release. In simple cases where semantic versioning is followed, minor releases may be ABI-compatible. Thus, in these simple cases, the major and minor releases correspond to the current and revision parameters used for -version-info.

 * configure.ac: get major and minor versions and substitute for Makefile
 * src/Makefile.am: Use major and minor versions for the -version-info current and revision parameters.