Cast away some MSVC warnings.
Odd. Visual Studio warns about a potential loss of data when passing a
`size_t` as an argument to a function taking an `unsigned long`.

I mean yes, I can see how that might happen _in principle, on a
different platform._  But why would `size_t` be wider than an
`unsigned long`?  And it can't be a sign issue, because that'd give me
a different warning.