Make `esc` write directly into `std::string`.
This is an optimisation we can use in C++17.  Instead of telling
`PQescapeStringConn` to write into a buffer, and then copying the text
from that buffer to a `std::string`, we can set up a `std::string` first
and tell `PQescapeStringConn` to do its work in there.  Then, shrink to
fit.

Subtle detail: we can't use `std::string::reserve` to allocate the
buffer.  It's not reliable, and we're not allowed to write beyond the
part that is actually used.