Make `zview` constructors `explicit`.
Since `zview` is effectively a marker type for "this string has a zero
at the end," we really only want it to be applied explicitly.  Otherwise
sooner or later we'll pass a `const char *` or a `std::string_view`
where a `zview` is expected, and fail to notice because of implicit
promotion.