Drop `buffer_budget`.
Now that `str` is gone, in most cases, we no longer absolutely need to
know buffer sizes for string representations at compile time.  (The one
exception seems to be with floating-point types.)

A `constexpr` variable was a very strict format for carring the
information, and it's nice to have guarantees.  But exactly because of
that, I couldn't always provide a meaningful `buffer_budget` anyway.  I
was already starting to toy with negative values to indicate uncertainty
at compile time, and that's just a little hairier than I like things.