Add clang-format config.
I'm going to try it.  I'm not completely happy with the formatting I can
get out of clang-format, but I've been wanting _something_ for so long.

My main remaining gripe is with the use of braces.  I have a simple rule
for those: feel free to leave out the braces on a single-statement block
_so long as matched if/else blocks are formatted identically._  That
does not seem to be something that clang-format can do.

The next-biggest gripe is with alignment and indentation:

    return std::vector<errorhandler *>{std::begin(m_errorhandlers),
                                       std::end(m_errorhandlers)};

That's just horrible.  I don't want that first argument on the same
line as the opening brace — if you've got a list of things that won't
fit the line, start breaking lines right _before_ the first item, so the
structure remains clear at a glance.  Nor do I like creative horizontal
alignment — indent or don't indent, but don't go padding to some
arbitrary width.

It seems that I can turn these horrors off in the configuration, but
only with parentheses, not with braces.  ☹