Less preprocessor.
Trying to minimise `#ifdef` code.  One reason is that the preprocessor
is supposed to be gradually going away (as it has been for decades…)
but another is that reduces the portion of the code which might contain
just about any syntax error or other obvious mistake, without generating
an error or warning on some particular system.

Not going as well as I'd hoped though.  I'm getting errors about the
inactive portion of an `if constexpr / else` clause calling undeclared
functions.  I thought the compiler was supposed to ignore that.