Support SQL arrays in `to_string` and friends.
Fixes #242.

Supports SQL arrays in `to_string` and friends.  This menas that you can
now pass a `std::vector` or `std::array` into `quote()`, and get your
container represented as an SQL array.

A nice next step after this would be to support `from_string` on SQL
arrays as well.  The one complication there is that we need to know the
client encoding in order to parse the SQL output correctly.