Provide connection string upon request.
Not quite as simple as it may seem, since libpq does not keep the exact
string anywhere that we can get it.  Nor did I want to store the string
inside a connection, since libpq does store the essentials and it would
incur unjustified extra weight.

Why do I want this feature in the first place?  I mean to allow
robusttransaction to "clone" a connection without introducing a factory
object into the API.  I toyed with giving `connection` a "clone" method,
but the name didn't sit quite right with me.  The metaphor does, but
users (of older libpqxx versions especially) might expect some current
connection state to be copied along.  Keeping the operation private to
its one use-case allows me to put off this decision, or perhaps avoid it
forever.