Rewrite robusttransaction using `txid_status`.
This requires postgres 10, but it's such a massive improvement that I
think that it's absolutely worth the pain.  The transaction log table
is cone, the fractal tree of awkward special cases is gone, basically
the server finally supports the functionality directly.  It came up in
discussions with the PostgreSQL core team a long time ago as I worked
on this functionality, but at the time the solution was still too far
in the future, and I didn't really touch the code for a while.

The new backend function simply tells you whether a transaction has
committed.  It's the "transaction log table" built into the backend
in a simple, convenient package, without the edge cases!  The handling
of the log table in robusttransaciton had come to a point where I did
not have great faith in its solidity in all cases.  That's all gone
now.

This does leave us with one incompatible change: if your code passes a
transaction log table name, that will no longer compile.  The
parameter is gone.  I could have continued to accept it for
compatibility, but I'd rather see immediate friction and cleanup than
have people continue to build client software based on an obsolete
mental model of how robusttransaction worked.