Use slightly lighter map for transaction statuses.
Replaces a `const std::map` with a `const std::unordered_map`, using a
trivial hash function which just reads the first character.  Happens to
be a perfect hash function for this case.  I toyed with switching on
that first character and then doing an extra check on "this is the
string I expect given this first character," but that gets really
tedious and brittle.  And it's really just a special case of the generic
facilities in the standard library.

Still, I'm not too happy about this.  I want my mapping `constexpr`!
Yes, I could easily do that with a simple array and a linear lookup
(no point getting clever for 3 entries), but it's back to being just a
little more tedious than what I've got here.  I think.