build: CMake pqxx improvements (#249)
* build: cmake: libpqxx: fix indentation format

Previously, 4 spaces were used for both indentation and line continuations. This uses 4 spaces for indentation and tabs for line continuations.

 * src/CMakeLists.txt{,.template}: Use tabs for line continuations

* build: cmake: libpqxx: remove separate shared and static library targets

Previously, there were separate shared and static library targets defined for libpqxx. This works alright on Linux, but on Windows the shared library interface (i.e. `libpqxx.lib`) which coincides with the actual dynamic runtime library (i.e. `libpqxx.dll`) will clobber the static library (i.e. `libpqxx.lib`). Instead, one library target should be defined, and the user will select which to build using `-DBUILD_SHARED_LIBS`. This is the recommended approach for CMake in any case.

 * src/CMakeLists.txt{,.template}: define a single library target for libpqxx
 * test/{,unit/}CMakeLists.txt{,.template}: use the pqxx target

* build: cmake: libpqxx: create and install library symlinks

This creates and installs a symlink for e.g. `libpqxx.so -> libpqxx-$major.$minor.so`.

 * src/CMakeLists.txt{,.template}: create and install library symlinks

* build: cmake: libpqxx: install pkg-config file

This installs the pkg-config file for libpqxx when using CMake, like with the autotools build.

 * libpqxx.pc.in: remove library and include directories for postgresql (private dependency)
 * src/CMakeLists.txt{,.template}: install pkg-config file with libpqxx