Factor out some string allocations.
I recently moved the `make_shared()` for query strings out of the
`result` class, and into the caller.  Not for the first time, I just
ran into a happy side effect: when the query is a constant string, I
can now create the `shared_ptr` once and re-use it.

Is this thread-safe?  As far as I can tell, yes it is, as of C++11:
 * Initialisation of static variables is serialised in C++11.
 * Once initialised, the `shared_ptr` is `const`.
 * Reference-counting is thread-safe.

My one remaining doubt is: could the compiler run the `make_shared`
concurrently?  But I guess even in that case, it would not leak the
`string` because the reference-counting is thread-safe.