Streamline COPY support.
Strip out a lot of overhead in how and where COPY was done.  We used to
need the overhead just because there were still two COPY implementations
(the old "table stream" and the new "stream to/from") which had to share
them.  But also, instead of tunneling calls through transaction and on
from there to connection, just go straight to connection.