East const.
Whew, that was a big job.  Change from `const char *` style to the
`char const *` style which experts recommend.  It's more consistent and
more "teachable."  It'll take some getting used to, for me as well as
for users, but if it makes our world better, I personally would prefer
to go through that pain now.

Of course that's a big "if," and we should never ignore the "if."  In
this case I've seen a few presentations make a strong case for the
change.