Test for nasty largeobject stream bug.
@tomlankhorst has discovered a nasty bug with `ilostream`: on systems
where `char` is signed, if a large object contains a byte value `0xff`
(i.e. -1 as a signed char) at a buffer boundary, that byte gets
mistaken for a -1, i.e. an end-of-file marker.

See #284 for bug details, and #286 for Tom's fix.  We'll briefly have
failing tests before I merge his fix.