Fix handling of eof() in streambuffer underflow.
@tomlankhorst has identified a serious bug which affected the
`ilostream` class.  If you opened a large object _as an ilostream,_
and it contained a byte with value `0xff` _at a buffer boundary,_
on a system where `char` is a signed type (such as x86-compatible
systems), then reading of the large object would **mistake the
`0xff` byte for end-of-file** and stop reading prematurely.