Streamline check for `PGresult` errors.
We were having some problems recognising and reporting errors from libpq
functions returning `PGresult`.  In particular, we'd query the
connection's encoding before reporting an error, at which point libpq
would check the connection's status, and cause an error if it was no
longer OK, which triggered an error and deprived us of the change to
report the error as we saw fit.

This may help with #280.  Certainly it will become a bit easier to
debug.

I removed `connection::check_result`, but this function was never called
or referenced from outside the library.  It was marked `PQXX_PRIVATE`,
which implies that there was no ABI impact from removing it.