Move row conversions.
Clean up result streaming a bit by moving the code to convert a row to a
tuple into `row`, and the handling of nulls into `from_string(field)`.

Next step though is to stop repeating, for every row in a streaming
result iteration, the check that the tuple and the row have the same
number of fields.  The compiler won't be able to aliminate those
repeated checks since the row size is hidden behind a libpq call.