Prefer `pg_config` over `pkg-config` for headers.
Works around #291.  In the long run though, we'll want to go with
`pkg-config` since `pg_config` is deprecated.

Also, the configure script will no longer fail in the case where neither
`pkg-config` nor `pg_config` is available, and the
`--with-postres-include` option was not specified, but the compiler can
find the postgres headers anyway.

Bumped VERSION to 7.0.6.