Unified the test suites.
Folds the tests in `test/test*.cxx` and `test/unit/test*.cxx` into a
single build.

Several benefits:
 * Less work & more parallelism during the build.
 * Convenience in running tests.
 * Scope for parallelism in test runs.
 * Easier to manage, e.g. to move the tests to better places.
 * Less templating.  CMake can do the globbing itself.
 * Increased scope for parallel test runs.

I tried implementing parallel test runs.  On a dual-core machine there
was no benefit whatsoever, but that was with severe limitations on what
tests could be run in parallel.  I can probably do better next time,
especially with more cores.  Yes yes yes, I know, the tests will run
fast anyway on a more modern machine, but they're still a linear part of
the build, making them a nice bottleneck to attack.

The bigger picture is that I'm trying to speed up the "broad-spectrum"
test script, `tools/test_all.py`.  For the builds themselves, I'm aiming
for parallelism within a single build, i.e. have all cores working on
one and the same build, with no other build going on in parallel.  That
requires optimal use of the CPUs during the build.  CMake with Ninja is
great for that, but with the `configure` builds, building two separasate
test suites is just a waste of wall-clock time.

Truth be told though, it looks as if the `configure` scripts themselves
are now the bottleneck on a fast multicore system!  I'm hoping to run
those in parallel as well, perhaps even concurrently with an actual build.